Move base_editor to folder utils
