+++ src/helpers/utils/base_editor.js
